review rating and registrations user

+++ flat/models/flat.py
@@ -12,10 +12,9 @@ class Flat(models.Model):
     cost = models.PositiveIntegerField()
     comfortable = models.CharField(max_length=200, null=True, blank=True)
     photos = models.ImageField(upload_to='flat_images', null=True, blank=True)
-
     max_guest = models.SmallIntegerField()
     arena_timeline = models.CharField(choices=ArenaTimeLine.choices(), max_length=200)
     total_area = models.SmallIntegerField()
     date_publisher = models.DateTimeField(auto_now=True)
     rating = GenericRelation('community.rating', null=True, blank=True)
-    avg_rating = models.IntegerField(default=0)
+    avg_rating = models.FloatField(default=0)
